refactored utils file
