+++ app/plugins/adp-backend/src/deliveryProgramme/deliveryProgrammeStore.ts
@@ -1,7 +1,7 @@
 import { Knex } from 'knex';
 import { NotFoundError } from '@backstage/errors';
 import { DeliveryProgramme } from '@internal/plugin-adp-common';
-import { createName } from '../utils';
+import { createName } from '../utils/utils';
 
 const TABLE_NAME = 'delivery_programme';
 type Row = {
