+++ app/plugins/adp-backend/src/providers/AlbTransformers.ts
@@ -1,6 +1,6 @@
 import { GroupEntity } from '@backstage/catalog-model';
 import { ArmsLengthBody } from '@internal/plugin-adp-common';
-import { createTransformerTitle } from '../utils';
+import { createTransformerTitle } from '../utils/utils';
 
 export type GroupTransformer = (
   armslenghBody: ArmsLengthBody,
