+++ app/plugins/adp-backend/src/utils/utils.test.ts
@@ -2,12 +2,10 @@ import {
   createName,
   createTransformerTitle,
   getCurrentUsername,
-  checkForDuplicateTitle,
-  getProgrammeManagerDetails,
+  checkForDuplicateTitle
 } from './utils';
 import express from 'express';
-import { NotFoundError } from '@backstage/errors';
-import { exampleCatalog } from './deliveryProgramme/programmeTestData';
+
 
 describe('createName', () => {
   it('replaces spaces with dashes and converts to lowercase', () => {
@@ -96,22 +94,3 @@ describe('getCurrentUsername', () => {
   });
 });
 
-describe('getProgrammeManagerDetails', () => {
-  it('returns the programme manager details', async () => {
-    await expect(
-      getProgrammeManagerDetails(
-        'a9dc2414-0626-43d2-993d-a53aac4d73421',
-        exampleCatalog,
-      ),
-    ).resolves.toEqual({ email: 'test1.test@onmicrosoft.com', name: 'Test1 Test' });
-  });
-
-  it('returns error if name is not found', async () => {
-    expect(
-      getProgrammeManagerDetails(
-        'a9dc2414-0626-43d2-993d-a53aac4d7341',
-        exampleCatalog,
-      ),
-    ).rejects.toThrow(NotFoundError);
-  });
-});
